refactor(pam-rdp): consolidate HandleConnection + cargo fmt
Moves the platform-agnostic HandleConnection method out of the
unix/windows cgo files into bridge_cgo_shared.go. Also applies
cargo fmt to bridge.rs and ffi.rs to satisfy CI.
