+++ packages/pam/handlers/rdp/bridge_cgo_shared.go
@@ -9,7 +9,49 @@ package rdp
 */
 import "C"
 
-import "fmt"
+import (
+	"context"
+	"errors"
+	"fmt"
+	"net"
+)
+
+func (p *RDPProxy) HandleConnection(ctx context.Context, clientConn net.Conn) error {
+	defer clientConn.Close()
+	if p.config.SessionLogger != nil {
+		defer func() {
+			_ = p.config.SessionLogger.Close()
+		}()
+	}
+
+	bridge, err := StartWithReadWriter(
+		clientConn,
+		p.config.TargetHost,
+		p.config.TargetPort,
+		p.config.InjectUsername,
+		p.config.InjectPassword,
+		p.config.AcceptorUsername,
+	)
+	if err != nil {
+		return fmt.Errorf("rdp proxy: start bridge: %w", err)
+	}
+	defer bridge.Close()
+
+	waitErr := make(chan error, 1)
+	go func() { waitErr <- bridge.Wait() }()
+
+	select {
+	case err := <-waitErr:
+		if err != nil && !errors.Is(err, ErrInvalidHandle) {
+			return fmt.Errorf("rdp proxy: session: %w", err)
+		}
+		return nil
+	case <-ctx.Done():
+		_ = bridge.Cancel()
+		<-waitErr
+		return ctx.Err()
+	}
+}
 
 // Wait blocks until the session ends. Idempotent.
 func (b *Bridge) Wait() error {
